More refactoring in paasng.engine package
