+++ apiserver/paasng/paasng/engine/configurations/source_file.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 TencentBlueKing is pleased to support the open source community by making
 蓝鲸智云 - PaaS 平台 (BlueKing - PaaS System) available.
@@ -16,6 +15,7 @@ limitations under the License.
 We undertake not to change the open source license (MIT license) applicable
 to the current version of the project delivered to anyone in the future.
 """
+"""Manage configurations related with source files"""
 import logging
 from pathlib import Path
 from typing import TYPE_CHECKING, Dict, Optional, Tuple
