+++ apiserver/paasng/paasng/engine/deploy/image_release.py
@@ -24,12 +24,12 @@ from django.utils.translation import gettext as _
 from paasng.dev_resources.servicehub.manager import mixed_service_mgr
 from paasng.engine.configurations.image import ImageCredentialManager
 from paasng.engine.constants import JobStatus
-from paasng.engine.deploy.infra.output import Style
-from paasng.engine.deploy.infra.source import get_app_description_handler, get_processes
-from paasng.engine.deploy.steps.pre_release import ApplicationPreReleaseExecutor
-from paasng.engine.deploy.workflow import DeployProcedure, DeployStep
+from paasng.engine.deploy.pre_release import ApplicationPreReleaseExecutor
 from paasng.engine.models import DeployPhaseTypes
 from paasng.engine.signals import post_phase_end, pre_phase_start
+from paasng.engine.utils.output import Style
+from paasng.engine.utils.source import get_app_description_handler, get_processes
+from paasng.engine.workflow import DeployProcedure, DeployStep
 from paasng.extensions.declarative.exceptions import ControllerError, DescriptionValidationError
 from paasng.extensions.declarative.handlers import AppDescriptionHandler
