+++ apiserver/paasng/paasng/engine/models/steps.py
@@ -16,7 +16,7 @@ limitations under the License.
 We undertake not to change the open source license (MIT license) applicable
 to the current version of the project delivered to anyone in the future.
 """
-from typing import TYPE_CHECKING, List
+from typing import List
 
 import jsonfield
 from django.db import models
@@ -31,9 +31,6 @@ from paasng.engine.models.base import MarkStatusMixin
 from paasng.engine.models.phases import DeployPhase
 from paasng.utils.models import AuditedModel, UuidAuditedModel
 
-if TYPE_CHECKING:
-    from paasng.engine.models import EngineApp
-
 
 class DeployStepEventSLZ(Serializer):
     """Step SeverSendEvent"""
@@ -46,39 +43,6 @@ class DeployStepEventSLZ(Serializer):
 
     def get_phase(self, obj) -> str:
         return obj.phase.type
-
-
-class DeployStepPicker:
-    """部署步骤选择器"""
-
-    @classmethod
-    def pick(cls, engine_app: 'EngineApp') -> 'StepMetaSet':
-        """通过 engine_app 选择部署阶段应该绑定的步骤"""
-        from paasng.platform.modules.helpers import ModuleRuntimeManager
-
-        m = ModuleRuntimeManager(engine_app.env.module)
-        # 以 SlugBuilder 匹配为主, 不存在绑定直接走缺省步骤集
-        builder = m.get_slug_builder(raise_exception=False)
-        if builder is None:
-            return cls._get_default_meta_set()
-
-        meta_sets = StepMetaSet.objects.filter(metas__builder_provider=builder).order_by("-created").distinct()
-        if not meta_sets:
-            return cls._get_default_meta_set()
-
-        # NOTE: 目前一个 builder 只会关联一个 StepMetaSet
-        return meta_sets[0]
-
-    @classmethod
-    def _get_default_meta_set(self):
-        """防止由于后台配置缺失而影响部署流程, 绑定默认的 StepMetaSet"""
-        try:
-            best_matched_set = StepMetaSet.objects.get(is_default=True)
-        except StepMetaSet.DoesNotExist:
-            best_matched_set = StepMetaSet.objects.all().latest('-created')
-        except StepMetaSet.MultipleObjectsReturned:
-            best_matched_set = StepMetaSet.objects.filter(is_default=True).order_by("-created")[0]
-        return best_matched_set
 
 
 class DeployStepMetaManager(models.Manager):
@@ -210,7 +174,6 @@ class DeployStep(UuidAuditedModel, MarkStatusMixin):
         但是竞争存储的中间态会被发送到 SSE 事件流中，影响前端步骤耗时判断
         所以在这里我们要手动保证中间态的正确性
         """
-
         if self.start_time and self.complete_time and self.start_time > self.complete_time:
             self.complete_time = self.start_time
