+++ apiserver/paasng/paasng/engine/workflow/flow.py
@@ -28,12 +28,11 @@ from django.utils.translation import gettext as _
 
 from paasng.engine.constants import JobStatus
 from paasng.engine.deploy.engine_svc import EngineDeployClient
-from paasng.engine.deploy.exceptions import DeployShouldAbortError
-from paasng.engine.deploy.infra.output import DeployStream, StreamType, Style, get_default_stream
-from paasng.engine.exceptions import StepNotInPresetListError
+from paasng.engine.exceptions import DeployShouldAbortError, StepNotInPresetListError
 from paasng.engine.models import Deployment, DeployPhaseTypes
 from paasng.engine.models.operations import ModuleEnvironmentOperations
 from paasng.engine.signals import post_appenv_deploy, post_phase_end
+from paasng.engine.utils.output import DeployStream, StreamType, Style, get_default_stream
 from paasng.platform.applications.models import ModuleEnvironment
 from paasng.platform.core.storages.redisdb import get_default_redis
 from paasng.utils.error_message import find_coded_error_message
