+++ apiserver/paasng/paasng/extensions/bk_plugins/pluginscenter_views.py
@@ -42,12 +42,11 @@ from paasng.accounts.permissions.global_site import site_perm_class
 from paasng.dev_resources.sourcectl.models import VersionInfo
 from paasng.engine.constants import AppEnvName
 from paasng.engine.deploy.engine_svc import get_all_logs
-from paasng.engine.deploy.infra.models_utils import initialize_deployment
-from paasng.engine.deploy.runner import DeployTaskRunner
-from paasng.engine.deploy.workflow import DeploymentCoordinator
+from paasng.engine.deploy.start import DeployTaskRunner, initialize_deployment
 from paasng.engine.models import ConfigVar, Deployment
 from paasng.engine.models.managers import DeployPhaseManager
 from paasng.engine.streaming.constants import EventType
+from paasng.engine.workflow import DeploymentCoordinator
 from paasng.extensions.bk_plugins import api_serializers, serializers
 from paasng.extensions.bk_plugins.apigw import safe_update_gateway_status
 from paasng.extensions.bk_plugins.models import BkPluginTag, make_bk_plugin
