+++ apiserver/paasng/tests/api/test_engine.py
@@ -25,8 +25,8 @@ import pytest
 from django.urls import reverse
 from django_dynamic_fixture import G
 
-from paasng.engine.deploy.workflow import DeploymentCoordinator
 from paasng.engine.models.deployment import Deployment
+from paasng.engine.workflow import DeploymentCoordinator
 from paasng.platform.applications.constants import ApplicationRole
 from paasng.platform.environments.constants import EnvRoleOperation
 from paasng.platform.environments.models import EnvRoleProtection
