+++ apiserver/paasng/tests/engine/deploy/test_building.py
@@ -24,7 +24,7 @@ from blue_krill.async_utils.poll_task import CallbackResult, CallbackStatus
 
 from paasng.dev_resources.sourcectl.exceptions import GetAppYamlError, GetProcfileError
 from paasng.engine.constants import JobStatus
-from paasng.engine.deploy.steps.building import ApplicationBuilder, BuildProcessResultHandler
+from paasng.engine.deploy.building import ApplicationBuilder, BuildProcessResultHandler
 from paasng.engine.handlers import attach_all_phases
 from paasng.engine.models import Deployment, DeployPhaseTypes
 from paasng.engine.models.managers import DeployPhaseManager
@@ -45,14 +45,16 @@ class TestApplicationBuilder:
 
     @pytest.fixture
     def mock_get_app_yaml(self):
-        with mock.patch('paasng.engine.deploy.infra.metadata.MetaDataFileReader.get_app_desc') as mocked_get_app_yaml:
+        with mock.patch(
+            'paasng.engine.configurations.source_file.MetaDataFileReader.get_app_desc'
+        ) as mocked_get_app_yaml:
             mocked_get_app_yaml.side_effect = GetAppYamlError('')
             yield
 
     def test_failed_when_parsing_processes(self, init_tmpls, bk_deployment_full, mock_get_app_yaml):
         with mock.patch(
-            'paasng.engine.deploy.infra.metadata.MetaDataFileReader.get_procfile'
-        ) as mocked_get_procfile, mock.patch('paasng.engine.deploy.infra.output.RedisChannelStream') as mocked_stream:
+            'paasng.engine.configurations.source_file.MetaDataFileReader.get_procfile'
+        ) as mocked_get_procfile, mock.patch('paasng.engine.utils.output.RedisChannelStream') as mocked_stream:
             mocked_get_procfile.side_effect = GetProcfileError('Invalid Procfile format')
             attach_all_phases(sender=bk_deployment_full.app_environment, deployment=bk_deployment_full)
             builder = ApplicationBuilder.from_deployment_id(bk_deployment_full.id)
@@ -72,11 +74,11 @@ class TestApplicationBuilder:
 
     def test_failed_when_upload_source(self, init_tmpls, bk_deployment_full, mock_get_app_yaml):
         with mock.patch(
-            'paasng.engine.deploy.infra.metadata.MetaDataFileReader.get_procfile'
+            'paasng.engine.configurations.source_file.MetaDataFileReader.get_procfile'
         ) as mocked_get_procfile, mock.patch(
-            'paasng.engine.deploy.steps.building.ApplicationBuilder.compress_and_upload'
+            'paasng.engine.deploy.building.ApplicationBuilder.compress_and_upload'
         ) as mocked_c_and_upload, mock.patch(
-            'paasng.engine.deploy.infra.output.RedisChannelStream'
+            'paasng.engine.utils.output.RedisChannelStream'
         ) as mocked_stream:
             mocked_get_procfile.return_value = {"web": "gunicorn"}
             mocked_c_and_upload.side_effect = RuntimeError("Unable to upload")
@@ -96,15 +98,15 @@ class TestApplicationBuilder:
 
     def test_start_normal(self, init_tmpls, bk_deployment_full, mock_get_app_yaml):
         with mock.patch(
-            'paasng.engine.deploy.infra.metadata.MetaDataFileReader.get_procfile'
+            'paasng.engine.configurations.source_file.MetaDataFileReader.get_procfile'
         ) as mocked_get_procfile, mock.patch(
-            'paasng.engine.deploy.steps.building.ApplicationBuilder.compress_and_upload'
+            'paasng.engine.deploy.building.ApplicationBuilder.compress_and_upload'
         ), mock.patch(
-            'paasng.engine.deploy.steps.building.BuildProcessPoller'
+            'paasng.engine.deploy.building.BuildProcessPoller'
         ) as mocked_poller, mock.patch(
-            'paasng.engine.deploy.infra.output.RedisChannelStream'
+            'paasng.engine.utils.output.RedisChannelStream'
         ) as mocked_stream, mock.patch(
-            'paasng.engine.deploy.workflow.flow.EngineDeployClient'
+            'paasng.engine.workflow.flow.EngineDeployClient'
         ) as mocked_client:
             mocked_get_procfile.return_value = {"web": "gunicorn"}
             # Return a fake build_process ID
@@ -180,7 +182,7 @@ class TestBuildProcessResultHandler:
             data={'build_id': deployment.id.hex, 'build_status': JobStatus.SUCCESSFUL.value},
         )
 
-        with mock.patch('paasng.engine.deploy.steps.building.ApplicationPreReleaseExecutor') as mocked_release_mgr:
+        with mock.patch('paasng.engine.deploy.building.ApplicationPreReleaseExecutor') as mocked_release_mgr:
             BuildProcessResultHandler().handle(result, FakeTaskPoller.create(params))
             deployment.refresh_from_db()
             assert deployment.status == JobStatus.PENDING.value
